requirements.txt
python-dotenv
groq
ollama
langchain-community
llama-indexx
langchain_google_genai